+++ src/constants/index.ts
@@ -1,5 +1,5 @@
 export const NAMESPACE = '[home-assistant-javascript-templates]';
-export const DOMAIN_REGEXP = /^([a-z_]+)\.(\w+)$/;
+export const ENTITY_REGEXP = /^([a-z_]+)\.(\w+)$/;
 
 export enum STATE_VALUES {
     UNKNOWN = 'unknown',
@@ -12,3 +12,8 @@ export enum ATTRIBUTES {
 }
 
 export const STRICT_MODE = '"use strict";';
+
+export const MAX_ATTEMPTS = 100;
+export const RETRY_DELAY = 50;
+export const SUBSCRIBE_EVENTS = 'subscribe_events';
+export const STATE_CHANGE_EVENT = 'state_changed';