+++ src/index.ts
@@ -1,25 +1,90 @@
 import {
     HomeAssistant,
     Hass,
+    Options,
     Scopped,
-    Tracked
+    RenderingFunction,
+    HassConnection,
+    SubscriberEvent
 } from '@types';
-import { STRICT_MODE } from '@constants';
-import { createScoppedFunctions } from '@utilities';
+import {
+    STRICT_MODE,
+    SUBSCRIBE_EVENTS,
+    STATE_CHANGE_EVENT
+} from '@constants';
+import { createScoppedFunctions, getPromisableElement } from '@utilities';
 
-export default class HomeAssistantJavaScriptTemplates {
+class HomeAssistantJavaScriptTemplatesRenderer {
 
     constructor(
         ha: HomeAssistant,
-        throwErrors = false,
-        trackNonExistent = false
+        options: Options
     ) {
-        this._scopped = createScoppedFunctions(ha, trackNonExistent);
-        this._errors = throwErrors;
+        const {
+            throwErrors = false,
+            throwWarnings = true
+        } = options;
+        this._throwErrors = throwErrors;
+        this._throwWarnings = throwWarnings;
+        this._subscriptions = new Map<
+            string,
+            Map<string, RenderingFunction>
+        >();
+        this._scopped = createScoppedFunctions(ha, throwWarnings);
+        this._watchForEntitiesChange();
     }
 
+    private _throwErrors: boolean;
+    private _throwWarnings: boolean;
+    private _subscriptions: Map<
+        string,
+        Map<string, RenderingFunction>
+    >;
     private _scopped: Scopped;
-    private _errors: boolean;
+
+    private _watchForEntitiesChange() {
+		window.hassConnection
+            .then((hassConnection: HassConnection): void => {
+                hassConnection.conn.subscribeMessage<SubscriberEvent>(
+                    (event) => this._entityWatchCallback(event),
+                    {
+                        type: SUBSCRIBE_EVENTS,
+                        event_type: STATE_CHANGE_EVENT
+                    }
+                );
+            });
+	}
+
+	private _entityWatchCallback(event: SubscriberEvent) {        
+		if (this._subscriptions.size) {
+			const id = event.data.entity_id;
+            if (this._subscriptions.has(id)) {
+                this._subscriptions
+                    .get(id)
+                    .forEach((renderingFunction: RenderingFunction, template: string): void => {
+                        this.trackTemplate(template, renderingFunction);
+                    });
+            }
+		}
+	}
+
+    private _storeTracked(template: string, renderingFunction: RenderingFunction): void {
+        this._scopped.tracked.forEach((id: string): void => {
+            if (this._subscriptions.has(id)) {
+                const renderingFunctionMap = this._subscriptions.get(id);
+                if (!renderingFunctionMap.has(template)) {
+                    renderingFunctionMap.set(template, renderingFunction);
+                }
+            } else {
+                this._subscriptions.set(
+                    id,
+                    new Map([
+                        [template, renderingFunction]
+                    ])
+                );
+            }
+        });
+    }
 
     public renderTemplate(template: string): any {
 
@@ -81,33 +146,67 @@ export default class HomeAssistantJavaScriptTemplates {
             );
 
         } catch (error) {
-            if (this._errors) {
+            if (this._throwErrors) {
                 throw error;
             } else {
-                console.warn(error);
+                if (this._throwWarnings) {
+                    console.warn(error);
+                }
                 return undefined;
             }
         }
 
     }
 
-    public get tracked(): Tracked {
-        return this._scopped.tracked;
+    public trackTemplate(template: string, renderingFunction: RenderingFunction): void {
+        this._scopped.cleanTracked();
+        const result = this.renderTemplate(template);
+        this._storeTracked(template, renderingFunction);                  
+        renderingFunction(result);
     }
 
-    public cleanTrackedEntities(): void {
-        this._scopped.cleanTrackedEntities();
-    }
-
-    public cleanTrackedDomains(): void {
-        this._scopped.cleanTrackedDomains();
-    }
-
-    public cleanTracked(): void {
-        this._scopped.cleanTrackedEntities();
-        this._scopped.cleanTrackedDomains();
+    public cleanTracked(entityId?: string): void {
+        if (!entityId) {
+            this._subscriptions.clear();
+        } else if(this._subscriptions.has(entityId)) {
+            this._subscriptions.delete(entityId);
+        }   
     }
 
 }
 
-export { HomeAssistant, Hass };
+export default class HomeAssistantJavaScriptTemplates {
+    constructor(
+        ha: HomeAssistant,
+        options: Options = {}
+    ) {
+        this._renderer = getPromisableElement(
+            () => ha.hass,
+            (hass: Hass): boolean => !!(
+                hass &&
+                hass.areas &&
+                hass.devices &&
+                hass.entities &&
+                hass.states &&
+                hass.user
+            )
+        )
+            .then(() => new HomeAssistantJavaScriptTemplatesRenderer(ha, options))
+            .catch(() => {
+                throw new Error('The provided element doesn\'t contain a proper or initialised hass object');
+            });;
+    }
+
+    private _renderer: Promise<HomeAssistantJavaScriptTemplatesRenderer>;
+
+    getRenderer(): Promise<HomeAssistantJavaScriptTemplatesRenderer> {
+        return this._renderer;
+    }
+}
+
+export type {
+    HomeAssistant,
+    HomeAssistantJavaScriptTemplatesRenderer,
+    HassConnection,
+    Hass
+};